refactor(MetadataBar): port to TS; add category names
